Move data intercept to useDataInternal
